% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_cid}
\alias{get_cid}
\title{Retrieve the Concept ID Associated with a Key}
\usage{
get_cid(dd, key)
}
\arguments{
\item{dd}{A \code{data.tree} object representing the data dictionary.}

\item{key}{A key to locate the corresponding concept ID.}
}
\value{
The concept ID associated with the given key. If no match is found, returns NULL.
}
\description{
This function returns the concept ID (cid) corresponding to a given key
within the data.tree object representing the data dictionary.
}
\examples{
# Assuming `dd` is a properly structured `data.tree` object
# and 'my_key' is a valid key in the tree:
path <- get_path_to_example_dictionary() # Replace with path to your dictionary
dd <- construct_dictionary_tree(path)
concept_id <- get_cid(dd, 'University of Chicago Medicine')
concept_id
}
