% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_key}
\alias{get_key}
\title{Retrieve the Key Associated with a Concept ID from a Data Tree}
\usage{
get_key(dd, cid)
}
\arguments{
\item{dd}{A \code{data.tree} object representing the data dictionary.}

\item{cid}{A concept ID as a string or numeric value, used to locate the specific node
within the data tree whose key is desired.}
}
\value{
The key associated with the given concept ID. If the concept ID does not
exist, returns NULL.
}
\description{
This function takes a \code{data.tree} object representing a data dictionary
and a concept ID (cid), returning the "key" associated with that specific
concept ID within the data tree.
}
\examples{
# Assuming `dd` is a properly structured `data.tree` object
# and '123' is a valid concept ID in the tree:
key_value <- get_key(dd, '123')
}
