% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_path_to_example_dictionary}
\alias{get_path_to_example_dictionary}
\title{Get Path to an example dictionary}
\usage{
get_path_to_example_dictionary()
}
\value{
string Path to a folder containing JSON files for a example dictionary
}
\description{
Get Path to an example dictionary
}
\examples{
path <- get_path_to_example_dictionary()
print(path)

}
