% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_var_name}
\alias{get_var_name}
\title{Retrieve Variable Name for a Question Concept}
\usage{
get_var_name(dd, cid_str)
}
\arguments{
\item{dd}{The data dictionary tree}

\item{cid_str}{A string containing the concept id for a question concept.}
}
\value{
The variable name for the question concept.
}
\description{
This function retrieves the variable name for a question concept, given a
string containing it's Concept ID. This is particularly useful for labeling
data from Connect's BigQuery tables.
}
\examples{
path <- get_path_to_example_dictionary() # Replace with path to your dictionary
dd <- construct_dictionary_tree(path)
var_name <- get_var_name(dd, "d_142654897_d_461488577") # Should return "RcrtES_Aware_v1r0_Email"
var_name

}
